Rename directory "variableTrees" to "variableSelectors"
