+++ packages/libs/eda/src/lib/core/components/variableSelectors/VariableTree.scss
