+++ packages/libs/eda/src/lib/core/components/variableSelectors/hooks.ts
