+++ packages/libs/eda/src/lib/core/components/visualizations/implementations/MosaicVisualization.tsx
@@ -68,7 +68,7 @@ import { createVisualizationPlugin } from '../VisualizationPlugin';
 import { LayoutOptions } from '../../layouts/types';
 import SingleSelect from '@veupathdb/coreui/lib/components/inputs/SingleSelect';
 import { useInputStyles } from '../inputStyles';
-import { ClearSelectionButton } from '../../variableTrees/VariableTreeDropdown';
+import { ClearSelectionButton } from '../../variableSelectors/VariableTreeDropdown';
 import { Tooltip } from '@veupathdb/components/lib/components/widgets/Tooltip';
 import Banner from '@veupathdb/coreui/lib/components/banners/Banner';
